+++ src/StatusBus/Failure/DeliveryFailureHandlerInterface.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * @copyright Copyright (c) 2020 - 2026 Communitales GmbH (https://www.communitales.com/)
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Communitales\Component\StatusBus\Failure;
+
+use Communitales\Component\StatusBus\Handler\StatusHandlerInterface;
+use Communitales\Component\StatusBus\StatusMessage;
+use Throwable;
+
+interface DeliveryFailureHandlerInterface
+{
+    public function handleFailure(
+        Throwable $exception,
+        StatusHandlerInterface $handler,
+        StatusMessage $message,
+    ): void;
+}